+++ app.go
@@ -2,13 +2,14 @@ package main
 
 import (
 	"bufio"
-	"errors"
 	"flag"
 	"os"
 	"strconv"
 	"strings"
 
 	"github.com/op/go-logging"
+
+	"github.com/distributeddesigns/milestone1/commands"
 )
 
 // Globals
@@ -18,70 +19,8 @@ var (
 	logLevel = flag.Int("loglevel", 4, "CRITICAL: 0, ERROR: 1, WARNING: 2, NOTICE: 3, INFO: 4, DEBUG: 5")
 )
 
-// All valid command names
-type commandType int
-
-// Command enum!
-const (
-	Add commandType = iota
-	Quote
-	Buy
-	CommitBuy
-	CancelBuy
-	Sell
-	CommitSell
-	CancelSell
-	SetBuyAmount
-	CancelSetBuy
-	SetBuyTrigger
-	SetSellAmount
-	SetSellTrigger
-	CancelSetSell
-	DisplaySummary
-	DumpLog
-)
-
-var commandNames = []string{
-	"ADD",
-	"QUOTE",
-	"BUY",
-	"COMMIT_BUY",
-	"CANCEL_BUY",
-	"SELL",
-	"COMMIT_SELL",
-	"CANCEL_SELL",
-	"SET_BUY_AMOUNT",
-	"CANCEL_SET_BUY",
-	"SET_BUY_TRIGGER",
-	"SET_SELL_AMOUNT",
-	"SET_SELL_TRIGGER",
-	"CANCEL_SET_SELL",
-	"DISPLAY_SUMMARY",
-	"DUMPLOG",
-}
-
-// String representation of the commandType enum
-func (c commandType) String() string {
-	return commandNames[c]
-}
-
-// Convert string -> commandType enum
-func toCommandType(cmd string) (commandType, error) {
-	for i, name := range commandNames {
-		if strings.EqualFold(name, cmd) {
-			return commandType(i), nil
-		}
-	}
-
-	return commandType(0), errors.New("Not a valid command type")
-}
-
-type command struct {
-	ID     int
-	name   commandType
-	userID string
-	args   []string
-}
+// I suck at namespacing and don't want to type commands.Command over and over
+type command commands.Command
 
 func main() {
 	flag.Parse()
@@ -171,12 +110,12 @@ func parseCommand(s string) command {
 	// Almost all commands will follow this format
 	// TODO: Deal with the final "DUMPLOG,./testLOG"
 	ID, _ := strconv.Atoi(parts[0])
-	name, _ := toCommandType(parts[1])
+	name, _ := commands.ToCommandType(parts[1])
 	parsed := command{
 		ID:     ID,
-		name:   name,
-		userID: parts[2],
-		args:   parts[3:],
+		Name:   name,
+		UserID: parts[2],
+		Args:   parts[3:],
 	}
 
 	log.Debugf("Parsed as: %+v", parsed)
@@ -185,12 +124,12 @@ func parseCommand(s string) command {
 }
 
 func executeCommand(cmd command) error {
-	switch cmd.name {
-	case Add:
+	switch cmd.Name {
+	case commands.Add:
 		// do ADD
 		break
 	default:
-		log.Noticef("Not implemented: %s", cmd.name)
+		log.Noticef("Not implemented: %s", cmd.Name)
 		return nil
 	}
